feat: create room command
