+++ src/presentation/commands/commands.ts
@@ -2,6 +2,7 @@ export enum Commands {
   CONNECTION = 'connection',
   USER_TOKEN = 'userToken',
   DISCONNECT = 'disconnect',
+  CREATE_ROOM = 'createRoom',
   JOIN_ROOM = 'joinRoom',
   LEAVE_ROOM = 'leaveRoom',
   SEND_MSG = 'sendMessage',
