+++ src/presentation/commands/connection/disconnect.ts
