+++ src/presentation/defineNamespaces.ts
@@ -1,5 +1,5 @@
 import { Server } from 'socket.io'
-import { connectionCommand } from './commands/connection'
+import { connectionCommand } from './commands/connection/connection'
 import { RoomService } from '../application/roomService'
 
 export async function registerCommands(io: Server) {
